requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
faiss-cpu==1.7.1.post2
fastapi==0.68.1
filelock==3.1.0
gdown==4.0.1
h11==0.12.0
huggingface-hub==0.0.17
idna==3.2
joblib==1.0.1
nltk==3.6.3
numpy==1.21.2
packaging==21.0
Pillow==8.3.2
pydantic==1.8.2
pyparsing==2.4.7
PySocks==1.7.1
PyYAML==5.4.1
regex==2021.9.24
requests==2.26.0
sacremoses==0.0.46
scikit-learn==1.0
scipy==1.7.1
sentence-transformers==2.0.0
sentencepiece==0.1.96
six==1.16.0
sklearn==0.0
starlette==0.14.2
threadpoolctl==2.2.0
tokenizers==0.10.3
torch==1.9.1
torchvision==0.10.1
tqdm==4.62.3
transformers==4.11.0
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
